requirements.txt
intersight
prettytable
python-dotenv
pyVmomi